+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: sort and filter notes
 - feat: page turn animation
 - feat: custom font
+- feat: search over book content
 - fix: can't open book with "'" in title
 - fix: book style #24, #28
 - fix: toc overflow